fix: scheduler time promotion
This was the weirdest bug in my life since now! The diff work loop
called `Scheduler.unsafe_promoteDeadline` until it reaches the
`performance.now` and then the diff loop started to work!
Also after leaving the loop in rest for a time the `performance.now()`
promoted and the next diff loop should reach it by the same way!